1.139.0. fix(seguridad): endurecimiento de configuracion, auditoria y revocacion de sesiones
Auditoria del backend y correcciones de los hallazgos criticos y altos.

Seguridad (critico)
- settings: SECRET_KEY obligatoria y de longitud minima con DEBUG=False;
  DEBUG pasa a False por defecto; ALLOWED_HOSTS validado sin comodin.
- settings: bloque completo de cabeceras de seguridad (HSTS, SSL redirect,
  cookies Secure/HttpOnly/SameSite, nosniff, X-Frame-Options DENY,
  CSRF_TRUSTED_ORIGINS). `manage.py check --deploy` queda sin avisos.
- El registro aplica AUTH_PASSWORD_VALIDATORS: antes solo exigia 8
  caracteres y admitia contrasenas que el cambio y el reset rechazaban.
- Cambiar o restablecer la contrasena y cambiar el email revocan todas
  las sesiones: blacklist de refresh tokens + marca credentials_changed_at
  que invalida los access tokens ya emitidos.
- Separacion staff/superusuario en el panel de administracion: crear
  administradores y conceder privilegios exige superusuario. Se protege
  ademas el ultimo superusuario activo.
- Nuevo registro de auditoria (tabla audit_log + logger dedicado) para
  logins, cambios de credenciales, 2FA y acciones de administracion.
- Bloqueo temporal por cuenta ante intentos fallidos de contrasena y de
  segundo factor, complementando el limite por IP.
- Sal especifica en el token de verificacion de email.

Correccion (alto)
- La URL de cache de Redis se construye parseando la URL: `rstrip("/0")`
  rompia `redis://host:6379` (formato de Railway/Upstash) dejando la
  cache apuntando a un puerto inexistente.
- Importes monetarios pasan de float a Decimal de extremo a extremo
  (serializers, DTOs y casos de uso), sin redondeo silencioso.
- El registro es atomico: alta y contrasena en una sola transaccion.
- Emails normalizados a minusculas en todos los flujos.
- Indices compuestos en market_data_snapshots, positions, trade_history
  y price_alerts; tareas de retencion para snapshots y auditoria.
- Se rechazan operaciones con fecha futura.

Operacion y calidad
- Manejador de excepciones global con contrato de error unico y
  request-id de correlacion; middleware y logging estructurado en JSON.
- Sondas separadas: /api/health/live/ (liveness, sin dependencias) y
  /api/health/ (readiness, 503 si una dependencia falla).
- Paginacion en el listado de usuarios del panel de administracion.
- Esquema OpenAPI 3 generado desde el codigo (/api/schema, /api/docs).
- Reparada la doble codificacion de caracteres en views.py, BOMs
  eliminados y .gitattributes que evita que vuelva a ocurrir.
- pytest con entorno determinista y umbral de cobertura.

+++ backend/src/core/application/dto/alerts_dto.py
@@ -3,6 +3,7 @@ alerts_dto.py — DTOs para el módulo de Alertas de Precio.
 """
 
 from dataclasses import dataclass
+from decimal import Decimal
 from typing import Optional
 
 
@@ -11,7 +12,7 @@ class CreateAlertInputDTO:
     """Datos para crear una nueva alerta de precio."""
     asset_symbol: str
     condition: str      # "ABOVE" | "BELOW"
-    threshold_price: float
+    threshold_price: Decimal
     notes: str = ""
